refactor: BigDecimal 타입을 Integer 타입으로 변경

+++ src/main/java/io/github/spartateam6/commercepaymentsystem/domain/order/dto/OrderCreateResponse.java
@@ -2,12 +2,10 @@ package io.github.spartateam6.commercepaymentsystem.domain.order.dto;
 
 import io.github.spartateam6.commercepaymentsystem.domain.order.entity.Order;
 
-import java.math.BigDecimal;
-
 public record OrderCreateResponse(
         Long orderId,
         String orderNumber,
-        BigDecimal totalAmount
+        Integer totalAmount
 ) {
     public static OrderCreateResponse from(Order order) {
         return new OrderCreateResponse(
