+++ src/main/java/io/github/spartateam6/commercepaymentsystem/domain/order/dto/OrderPreviewResponse.java
@@ -1,19 +1,18 @@
 package io.github.spartateam6.commercepaymentsystem.domain.order.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderPreviewResponse(
         List<PreviewItem> items,
-        BigDecimal totalAmount
+        Integer totalAmount
 ) {
     public record PreviewItem(
             Long cartItemId,
             Long productId,
             String productName,
-            BigDecimal unitPrice,
+            Integer unitPrice,
             Integer quantity,
-            BigDecimal lineAmount
+            Integer lineAmount
     ) {
     }
 }
