+++ src/main/java/io/github/spartateam6/commercepaymentsystem/domain/order/service/OrderService.java
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -61,7 +60,7 @@ public class OrderService {
     public record CreateOrderItem(
             Product product,
             String productName,
-            BigDecimal unitPrice,
+            Integer unitPrice,
             Integer quantity
     ) {
     }
